LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-getPropertyType-expected.txt


FAIL sanity check trustedTypes.getPropertyType for the HTML script element. assert_equals: expected (string) "TrustedScript" but got (object) null
FAIL sanity check trustedTypes.getAttributeType. assert_equals: expected (string) "TrustedScript" but got (object) null
FAIL sanity check trustedTypes.getTypeMapping trustedTypes.getTypeMapping is not a function. (In 'trustedTypes.getTypeMapping()', 'trustedTypes.getTypeMapping' is undefined)
FAIL getPropertyType tests adapted from w3c/trusted-types polyfill assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL getAttributeType tests adapted from w3c/trusted-types polyfill assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL getTypeMapping tests adapted from w3c/trusted-types polyfill trustedTypes.getTypeMapping is not a function. (In 'trustedTypes.getTypeMapping()', 'trustedTypes.getTypeMapping' is undefined)
FAIL object[codeBase] is defined assert_true: expected true got false
FAIL object.codeBase is maybe defined assert_equals: expected true but got false
FAIL OBJECT[codeBase] is defined assert_true: expected true got false
FAIL OBJECT.codeBase is maybe defined assert_equals: expected true but got false
FAIL oBjEcT[codeBase] is defined assert_true: expected true got false
FAIL oBjEcT.codeBase is maybe defined assert_equals: expected true but got false
FAIL object[CODEBASE] is defined assert_true: expected true got false
PASS object.CODEBASE is maybe defined
FAIL OBJECT[CODEBASE] is defined assert_true: expected true got false
PASS OBJECT.CODEBASE is maybe defined
FAIL oBjEcT[CODEBASE] is defined assert_true: expected true got false
PASS oBjEcT.CODEBASE is maybe defined
FAIL object[codebase] is defined assert_true: expected true got false
PASS object.codebase is maybe defined
FAIL OBJECT[codebase] is defined assert_true: expected true got false
PASS OBJECT.codebase is maybe defined
FAIL oBjEcT[codebase] is defined assert_true: expected true got false
PASS oBjEcT.codebase is maybe defined
